refactor(live): extract types, helpers, renderers from Live.tsx
Live.tsx split from 2454 → 765 lines into 3 focused modules:
- liveTypes.ts (107 lines): interfaces, type guards, constants
- liveHelpers.ts (638 lines): pure functions, aggregation, parsers
- LiveRenderers.tsx (1018 lines): all React rendering components

Zero behavior change — pure file split for reuse by Activity tab.
Prepares for useAgentActivity hook and AgentToolStream (phases 3-5).
